refactor: ♻️ Moved error/service-result utilities into `utils/errors/` subfolder 🔀 Updated 26 import paths across routes, services, DAL, and components 🐛 Added D1 HTTP API non-ok response check before parsing JSON 🎨 Fixed `starting:opacity-50` missing `transition-opacity` on about page 🧪 Relocated to-action-result test with updated imports
