+++ app/features/host/services/income.server.ts
@@ -15,7 +15,7 @@ import type {
 import { toPagination } from "~/features/pagination/utils/to-pagination.server";
 import type { Prettify } from "~/types";
 import type { UUIDv7 } from "~/types/ids.server";
-import { tryCatch } from "~/utils/try-catch.server";
+import { tryCatch } from "~/utils/errors/try-catch.server";
 
 export type HostPaginatedPageParams = Prettify<
   BasePaginationParams & SortObject
