+++ app/features/host/services/transfers.server.ts
@@ -11,7 +11,7 @@ import {
 } from "~/features/host/utils/resolve-chart-context.server";
 import { toPagination } from "~/features/pagination/utils/to-pagination.server";
 import type { UUIDv7 } from "~/types/ids.server";
-import { tryCatch } from "~/utils/try-catch.server";
+import { tryCatch } from "~/utils/errors/try-catch.server";
 
 export async function loadTransfersPage(
   db: AppDb,
