+++ app/features/vans/services/van-detail.server.ts
@@ -1,6 +1,6 @@
 import type { AppDb } from "~/db/client.server";
 import { getVanBySlug } from "~/features/vans/dal/van.server";
-import { tryCatch } from "~/utils/try-catch.server";
+import { tryCatch } from "~/utils/errors/try-catch.server";
 
 export function loadVanBySlug(db: AppDb, slug: string) {
   return tryCatch(() => getVanBySlug(db, slug));
