+++ app/utils/errors/domain-error.server.ts
@@ -1,7 +1,7 @@
 /** biome-ignore-all lint/suspicious/noUnnecessaryConditions: DomainErrorCode cases reachable; biome mis-narrows after isDomainError */
 import { data } from "react-router";
-import { notFound } from "~/utils/not-found";
-import { err, type ServiceFailure } from "~/utils/service-result.server";
+import { notFound } from "~/utils/errors/not-found";
+import { err, type ServiceFailure } from "~/utils/errors/service-result.server";
 
 type DomainErrorCode = "NOT_FOUND" | "CONFLICT" | "INVALID_ID";
