+++ app/utils/errors/get-collection-state.ts
